+++ src/utils/file/config_reader.rs
@@ -36,7 +36,7 @@ pub fn read_mappings(args_mapping: Option<String>, cfg: &mut Config) -> Result<O
 pub fn read_api_proxy_config(args_api_proxy_config: Option<String>, cfg: &mut Config) -> Option<String> {
     let api_proxy_config_file: String = args_api_proxy_config.unwrap_or_else(|| file_utils::get_default_api_proxy_config_path(cfg.t_config_path.as_str()));
     api_proxy_config_file.clone_into(&mut cfg.t_api_proxy_file_path);
-    let api_proxy_config = read_api_proxy(api_proxy_config_file.as_str(), true);
+    let api_proxy_config = read_api_proxy(cfg, api_proxy_config_file.as_str(), true);
     match api_proxy_config {
         None => {
             warn!("cant read api_proxy_config file: {}", api_proxy_config_file.as_str());
@@ -90,12 +90,12 @@ pub fn read_mapping(mapping_file: &str) -> Result<Option<Mappings>, M3uFilterErr
     Ok(None)
 }
 
-pub fn read_api_proxy(api_proxy_file: &str, resolve_var: bool) -> Option<ApiProxyConfig> {
+pub fn read_api_proxy(config: &Config, api_proxy_file: &str, resolve_var: bool) -> Option<ApiProxyConfig> {
     file_utils::open_file(&std::path::PathBuf::from(api_proxy_file)).map_or(None, |file| {
             let mapping: Result<ApiProxyConfig, _> = serde_yaml::from_reader(file);
             match mapping {
                 Ok(mut result) => {
-                    match result.prepare(resolve_var) {
+                    match result.prepare(config, resolve_var) {
                         Err(err) => {
                             exit!("cant read api-proxy-config file: {}", err);
                         }